+++ Assets/Scripts/TES/Materials/LightweightMaterial.cs
@@ -8,14 +8,18 @@ namespace TESUnity
     /// </summary>
     public class LightweightMaterial : MWBaseMaterial
     {
+        private Material m_StandardPBR;
+        private Material m_StandardSimple;
         private Material m_CutoutPBRMaterial;
-        private Material m_CutoutMaterial;
+        private Material m_CutoutSimpleMaterial;
 
         public LightweightMaterial(TextureManager textureManager)
             : base(textureManager)
         {
-            m_CutoutPBRMaterial = Resources.Load<Material>("Materials/Lightweight-Default-PBR-Cutout");
-            m_CutoutMaterial = Resources.Load<Material>("Materials/Lightweight-Default-Cutout");
+            m_StandardPBR = Resources.Load<Material>("Materials/Lightweight-PBR");
+            m_StandardSimple = Resources.Load<Material>("Materials/Lightweight-Simple");
+            m_CutoutPBRMaterial = Resources.Load<Material>("Materials/Lightweight-PBR-Cutout");
+            m_CutoutSimpleMaterial = Resources.Load<Material>("Materials/Lightweight-Simple-Cutout");
         }
 
         public override Material BuildMaterialFromProperties(MWMaterialProps mp)
@@ -53,10 +57,10 @@ namespace TESUnity
 
         public override Material BuildMaterial()
         {
-            if (TESUnity.instance.materialType == TESUnity.MWMaterialType.PBR)
-                return new Material(Shader.Find("LightweightPipeline/Standard (Physically Based)"));
-            else
-                return new Material(Shader.Find("LightweightPipeline/Standard (Simple Lighting)"));
+            var pbr = TESUnity.instance.materialType == TESUnity.MWMaterialType.PBR;
+            var material = new Material(Shader.Find(string.Format("LightweightPipeline/Standard ({0})", (pbr ? "Physically Based" : "Simple Lighting"))));
+            material.CopyPropertiesFromMaterial(pbr ? m_StandardPBR : m_StandardSimple);
+            return material;
         }
 
         public override Material BuildMaterialBlended(ur.BlendMode sourceBlendMode, ur.BlendMode destinationBlendMode)
@@ -71,8 +75,7 @@ namespace TESUnity
         {
             var material = BuildMaterial();
             var pbr = TESUnity.instance.materialType == TESUnity.MWMaterialType.PBR;
-            material.CopyPropertiesFromMaterial(pbr ? m_CutoutPBRMaterial : m_CutoutMaterial);
-
+            material.CopyPropertiesFromMaterial(pbr ? m_CutoutPBRMaterial : m_CutoutSimpleMaterial);
             material.SetFloat("_Cutout", cutoff);
             return material;
         }