+++ Assets/Scripts/TES/NIF/NIFObjectBuilder.cs
@@ -228,7 +228,8 @@ namespace TESUnity
 					meshRenderer.enabled = false;
 				}
 
-                obj.isStatic = true;
+                if (TESUnity.instance.useStaticBatching)
+                    obj.isStatic = true;
 			}
 			
 			if(collidable)